requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
email-validator==2.2.0
